Add get_baudrate function and associated dependencies
Turns out that although termios.h is real useful, it has downfalls in the form of non-castable baud rate constants.  Conversion functions have also been written in this commit and are publicly declared.

Remember that until stated otherwise, nothing in this UART code has been tested or even attempted to be compiled!

Signed-off-by: Misha Turnbull <mishaturnbull@gmail.com>